github-release: rename __testutil__
